requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
astroid==3.0.1
attrs==23.1.0
autopep8==2.0.4
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
cloudinary==1.38.0
crispy-bootstrap4==2023.1
dill==0.3.7
Django==4.2.7
django-admin-interface==0.27.0
django-browser-reload==1.12.0
django-ckeditor==6.7.0
django-cloudinary-storage==0.3.0
django-colorfield==0.11.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-extensions==3.2.3
django-filter==23.4
django-js-asset==2.2.0
django-phonenumber-field==7.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
Faker==20.1.0
fontawesomefree==6.4.2
frozenlist==1.4.0
h11==0.14.0
idna==3.6
isort==5.12.0
mccabe==0.7.0
multidict==6.0.4
outcome==1.3.0.post0
packaging==23.2
phonenumbers==8.13.24
Pillow==10.1.0
platformdirs==3.11.0
psycopg2==2.9.9
pycodestyle==2.11.1
PyJWT==2.8.0
pylint==3.0.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3.post1
requests==2.31.0
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
text-unidecode==1.3
tomlkit==0.12.1
trio==0.23.1
trio-websocket==0.11.1
urllib3==2.1.0
webdriver-manager==4.0.1
wsproto==1.2.0
yarl==1.9.4
